Renamed files to fit style guide
